requirements.txt
matplotlib==3.7.2
mmsegmentation==1.1.2
numpy==1.23.5
nuscenes_devkit==1.1.10
open3d==0.16.0
opencv_python==4.7.0.72
Pillow==10.2.0
pygmtools==0.4.1
pyquaternion==0.9.9
scikit_learn==1.3.0
scipy==1.10.0
